Normalizing File Names Conforming to the electron standard file naming scheme
